merge w/ Servuo master

+++ Scripts/Services/PointsSystems/CasinoData.cs
@@ -12,7 +12,7 @@ namespace Server.Engines.Points
 		public override PointsType Loyalty { get { return PointsType.CasinoData; } }
 		public override TextDefinition Name { get { return m_Name; } }
 		public override bool AutoAdd { get { return true; } }
-		public override double MaxPoints { get { return int.MaxValue; } }
+		public override double MaxPoints { get { return double.MaxValue; } }
         public override bool ShowOnLoyaltyGump { get { return false; } }
 
         public static readonly int ChipCost = 100;
